+++ templates/types/streaming/nextjs/app/components/header.tsx
@@ -12,13 +12,12 @@ export default function Header() {
           href="https://www.llamaindex.ai/"
           className="flex items-center justify-center font-nunito text-lg font-bold gap-2"
         >
-          <span>Built by LlamaIndex</span>
+          <span>Built by</span>
           <Image
-            className="rounded-xl"
-            src="/llama.png"
-            alt="Llama Logo"
-            width={40}
-            height={40}
+            src="/logo.png"
+            alt="T-Systems Logo"
+            width={52}
+            height={62}
             priority
           />
         </a>